* ext/openssl/ossl_hmac.c (ossl_hmac_hexdigest, ossl_hmac_s_hexdigest),   ext/openssl/ossl_pkey_ec.c (ossl_ec_group_set_seed),   ext/openssl/ossl_ssl_session.c (ossl_ssl_session_to_der),   ext/openssl/ossl_pkcs7.c (numberof): suppress warnings.   [ruby-core:31932]
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@29346 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/openssl/ossl_pkcs7.c
@@ -61,7 +61,7 @@
     GetPKCS7ri(obj, p7ri); \
 } while (0)
 
-#define numberof(ary) (sizeof(ary)/sizeof(ary[0]))
+#define numberof(ary) (int)(sizeof(ary)/sizeof(ary[0]))
 
 #define ossl_pkcs7_set_data(o,v)       rb_iv_set((o), "@data", (v))
 #define ossl_pkcs7_get_data(o)         rb_iv_get((o), "@data")
